Brightspace solution as a separate branch
This needs to be sorted out.

Quoting from brightspace:
“”
The solution module contains part of the solution of the exercises that
you are meant to do.
This code was not made such that it would be distributed so be aware
that it has not been polished. We provide it such that all of you have
the same information. So, use it carefully not all might be exactly
what you need to implement.

For example, notice that the shading model implementation does not
implement the Phong Model but the Blinn Model.
I have not explained the Blinn model in the lecture and actually it was
not meant to be provided with the framework in the first place. The
Blinn model would be seen rather as an extension of the Phong Model
described in class.
The Blinn model is a “simplified” model of the Phong model which is
computationally more efficient specially with some specific light
conditions, and avoids problems that the Phong model has.
“”